wip: Merge file server and store; point /f to /f/
